requirements/prod.txt
# ============================================
# PRODUCCIÓN - PROJECT BRAIN
# ============================================

# Dependencias base
-r base.txt

# Dependencias ML
-r ml.txt

# Bases de datos
-r databases.txt

# Agentes
-r agents.txt

# API
-r api.txt


# =====================
# Runtime & Performance
# =====================
psutil>=5.9.8,<6.0.0
gunicorn>=21.2.0,<22.0.0
uvicorn[standard]>=0.24.0,<0.25.0


# =====================
# Observabilidad
# =====================
prometheus-client>=0.19.0,<1.0.0

opentelemetry-api>=1.22.0,<2.0.0
opentelemetry-sdk>=1.22.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.43b0,<0.44.0
opentelemetry-instrumentation-sqlalchemy>=0.43b0,<0.44.0
opentelemetry-instrumentation-redis>=0.43b0,<0.44.0


# =====================
# Logging
# =====================
structlog>=23.2.0,<24.0.0
python-json-logger>=2.0.7,<3.0.0
loguru>=0.7.2,<1.0.0


# =====================
# Seguridad
# =====================
cryptography>=41.0.7,<42.0.0
bcrypt>=4.1.2,<5.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
slowapi>=0.1.9,<0.2.0
limits>=3.6.0,<4.0.0


# =====================
# Cache & Background
# =====================
redis>=5.0.1,<6.0.0

celery>=5.3.6,<6.0.0
flower>=2.0.1,<3.0.0

arq>=0.27.1,<0.28.0
dramatiq>=1.16.0,<2.0.0


# =====================
# Compresión & Serialización
# =====================
brotli>=1.1.0,<2.0.0
zstandard>=0.22.0,<1.0.0

orjson>=3.9.15,<4.0.0
msgpack>=1.0.7,<2.0.0
ujson>=5.9.0,<6.0.0


# =====================
# Bases de datos
# =====================
psycopg2-binary>=2.9.9,<3.0.0
asyncpg>=0.29.0,<0.30.0
sqlalchemy>=2.0.23,<3.0.0
alembic>=1.13.1,<2.0.0

neo4j>=5.15.0,<6.0.0
chromadb>=0.4.22,<0.5.0


# =====================
# Modelos & Agentes
# =====================
langchain>=0.1.0,<0.2.0
openai>=1.6.0,<2.0.0
anthropic>=0.8.0,<0.9.0


# =====================
# API
# =====================
fastapi>=0.104.1,<0.110.0
websockets>=12.0,<13.0


# =====================
# Validación & Config
# =====================
pydantic>=2.5.2,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0,<3.0.0

python-dotenv>=1.0.1,<2.0.0
pyyaml>=6.0.1,<7.0.0


# =====================
# Sistema
# =====================
watchfiles>=0.21.0,<1.0.0
filelock>=3.13.1,<4.0.0
platformdirs>=4.1.0,<5.0.0

pendulum>=2.1.2,<3.0.0
pytz>=2023.4,<2025.0
babel>=2.14.0,<3.0.0


# =====================
# CLI & UX
# =====================
typer>=0.9.0,<1.0.0
rich>=13.7.0,<14.0.0


# =====================
# Async & HTTP
# =====================
anyio>=3.7.1,<4.0.0
httpx>=0.26.0,<0.27.0
aiohttp>=3.9.3,<4.0.0
aiofiles>=23.2.1,<24.0.0


# =====================
# Ciencia de datos (runtime)
# =====================
numpy>=1.24.4,<2.0.0
pandas>=2.1.4,<3.0.0
scipy>=1.11.4,<2.0.0


# =====================
# Grafos & Código
# =====================
networkx>=3.2.1,<4.0.0
tree-sitter>=0.21.0,<1.0.0
tree-sitter-languages>=1.8.0,<2.0.0
